fix: not interpreting the bytes of the numbers correctly
Bytes were being parsed as string, so the integer 0x1 will be converted into a 1 in ascci. Now 0x1 is stored as 0x1 in the string.

Also changed the return of data(), now being a std::vector<char> for better binary data managment.

Finally, now there will be different kinds of objects and  the duck, the ammo, all objects with a position are Object2Ds